pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "python-practice"
version = "0.1.0"
description = ""
authors = [
    {name = "JGMEYER",email = "meyer.justian@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "pygame-gui (>=0.6.14,<0.7.0)",
    "cairosvg (>=2.8.2,<3.0.0)",
    "pygame-ce (>=2.5.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
